pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-anonymous-chat-bot"
version = "0.1.0"
description = "Production-ready Telegram anonymous stranger chat bot."
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
  "aiogram>=3.4,<4.0",
  "alembic>=1.13,<2.0",
  "asyncpg>=0.29,<1.0",
  "fastapi>=0.110,<1.0",
  "greenlet>=3.0,<4.0",
  "httpx>=0.27,<1.0",
  "pydantic-settings>=2.2,<3.0",
  "redis[hiredis]>=5.0,<6.0",
  "sqlalchemy[asyncio]>=2.0.29,<3.0",
  "structlog>=24.1,<25.0",
  "uvicorn[standard]>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20,<1.0",
  "pytest>=8.1,<9.0",
  "pytest-asyncio>=0.23,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

